Separate components in to sub-folders
