+++ ui/src/components/tools/SystemInfo.vue
