+++ ui/src/layouts/MainLayout.vue
@@ -63,8 +63,8 @@ import MenuItems from 'components/MenuItems.vue'
 import menuList from '../components/styles/menuList'
 import { qHeaderStyle } from '../components/styles/qStyles'
 import { useQuasar } from 'quasar'
-import Reboot from 'src/components/Reboot.vue'
-import Shutdown from 'src/components/Shutdown.vue'
+import Reboot from 'src/components/system/Reboot.vue'
+import Shutdown from 'src/components/system/Shutdown.vue'
 import { defineComponent, ref, watch } from 'vue'
 import { useI18n } from 'vue-i18n'
 import type { QuasarLanguage } from 'quasar'
